+++ lib/src/models/current_weather_response.dart
@@ -5,6 +5,7 @@ class CurrentWeatherResponse {
   /// The geographical coordinates of the location.
   final Coordinate coordinates;
 
+  /// The main weather data, including temperature, humidity, and pressure.
   final WeatherData weatherData;
 
   /// The time of sunrise at the location.
@@ -26,16 +27,23 @@ class CurrentWeatherResponse {
   final String dataSource;
 
   /// Creates an instance of [CurrentWeatherResponse] from a map containing the response data.
+  ///
+  /// This factory constructor takes a map of raw data from the OpenWeather API
+  /// and parses it to create an instance of [CurrentWeatherResponse].
   factory CurrentWeatherResponse.fromMap(Map<String, dynamic> map) {
     return CurrentWeatherResponse(
       coordinates: Coordinate(
-          latitude: map['coord']['lat'], longitude: map['coord']['lon']),
-      sunrise: DateTime.fromMillisecondsSinceEpoch(map['sys']['sunrise'] * 1000,
-              isUtc: true)
-          .add(Duration(seconds: map['timezone'])),
-      sunset: DateTime.fromMillisecondsSinceEpoch(map['sys']['sunset'] * 1000,
-              isUtc: true)
-          .add(Duration(seconds: map['timezone'])),
+        latitude: map['coord']['lat'],
+        longitude: map['coord']['lon'],
+      ),
+      sunrise: DateTime.fromMillisecondsSinceEpoch(
+        map['sys']['sunrise'] * 1000,
+        isUtc: true,
+      ).add(Duration(seconds: map['timezone'])),
+      sunset: DateTime.fromMillisecondsSinceEpoch(
+        map['sys']['sunset'] * 1000,
+        isUtc: true,
+      ).add(Duration(seconds: map['timezone'])),
       cityName: map['name'],
       countryCode: map['sys']['country'],
       timezoneOffset: Duration(seconds: map['timezone']),
@@ -45,21 +53,30 @@ class CurrentWeatherResponse {
   }
 
   /// Creates an instance of [CurrentWeatherResponse] with the given parameters.
-  CurrentWeatherResponse(
-      {required this.coordinates,
-      required this.weatherData,
-      required this.sunrise,
-      required this.sunset,
-      required this.cityName,
-      required this.countryCode,
-      required this.timezoneOffset,
-      required this.dataSource});
+  ///
+  /// This constructor is used to create an instance of [CurrentWeatherResponse]
+  /// with specified values for each of the properties.
+  CurrentWeatherResponse({
+    required this.coordinates,
+    required this.weatherData,
+    required this.sunrise,
+    required this.sunset,
+    required this.cityName,
+    required this.countryCode,
+    required this.timezoneOffset,
+    required this.dataSource,
+  });
 
   /// Returns a string representation of the weather response.
+  ///
+  /// The [toString] method provides a human-readable representation of the
+  /// [CurrentWeatherResponse] object, which includes the coordinates,
+  /// weather data, sunrise and sunset times, city and country name,
+  /// timezone offset, and data source.
   @override
   String toString() {
-    return 'WeatherResponse{coordinates: $coordinates, '
-        'weatherData: $weatherData'
+    return 'CurrentWeatherResponse{coordinates: $coordinates, '
+        'weatherData: $weatherData, '
         'sunrise: $sunrise, '
         'sunset: $sunset, cityName: $cityName, countryCode: $countryCode, '
         'timezoneOffset: $timezoneOffset, dataSource: $dataSource}';