+++ CHANGELOG.md
@@ -1,3 +1,11 @@
-## 0.0.1
+# Changelog
 
-* TODO: Describe initial release.
+## [1.0.0] - 2024-06-26
+
+### Added
+- Initial release of the `weather_manager` package.
+- Implemented `WeatherMan` class for fetching current weather and hourly forecast data.
+- Added support for fetching current weather data by latitude and longitude.
+- Added support for fetching hourly weather forecast data for up to 5 days.
+- Defined response classes: `CurrentWeatherResponse`, `ForecastWeatherResponse`, and `WeatherData`.
+- Included exception handling with `APIException` class.